+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Editor: Shorten long descriptions of foreign keys
 ### Plugins
 - Driver plugins: Driver::disconnect() to forget the logged user
+- Plugin igdb: Log in by the client secret instead of an access token obtained manually
 - New plugin: Display the first string column of the referenced row instead of a foreign key value, same as in Adminer Editor
 
 ## Adminer 6.1.0 (released 2026-09-14)